+++ src/plone/jsonapi/routes/zcml.py
@@ -17,20 +17,7 @@ class IConfig(Interface):
         """,
         required=False)
 
-    use_advanced_query = Bool(
-        title=u"Search uses Advanced Query",
-        description=u"""
-        Use Advanced Query if installed for catalog searches
-        """,
-        required=False)
 
-
-def configDirective(_context, register_api_routes=True, use_advanced_query=False):
-
-    if use_advanced_query:
-        from plone.jsonapi.routes import query
-        logger.info("*** USE ADVANCED QUERY ***")
-        query.USE_ADVANCED_QUERY = use_advanced_query
-
+def configDirective(_context, register_api_routes=True):
     import plone.jsonapi.routes
     plone.jsonapi.routes.REGISTER_API_ROUTES = register_api_routes